move utils to helper file
